Changed Patent. Added a test for the presence of a mainClassification and handling of assignees as an array.
Added AssigneeCountMapReduce to map assignees and reduce to the occurrences of each assignee. This is not working as expected, and returning all blanks.

Changed hive.txt to retrieve the patents with a main class of 428.

Changed PatentAbstractXmlMapReduceTest to add testing for main classifications.

Changed hive.txt to add commands related to main classifications.

Added PatentTest to test on whether it contains a main classification.

Added AssigneeCountMapReduceTest. It worked once in debug, but now is asking for Hadoop home.

Added SplitAssigneeFile in test\util. This is a one-off that reads a patent with a concatenated list of assignees and then writes one assignee at a time. This one works and it is how I am getting the working assignees.txt file.

Changed POM to version 1.5